+++ CHANGELOG.md
@@ -8,7 +8,7 @@ All notable changes to this project will be documented in this file. Dates are d
 
 > 13 May 2020
 
-- [https://github.com/doesdev/rollup-plugin-analyzer/pull/17](Add TS declarations)
+- [Add TS declarations](https://github.com/doesdev/rollup-plugin-analyzer/pull/17)
 - Update lockfile
 
 ### [3.2.2](https://github.com/doesdev/rollup-plugin-analyzer/compare/3.2.1...3.2.2)